+++ core/editable.js
@@ -29,6 +29,21 @@
 			this.editor = editor;
 
 			/**
+			 * Indicates editable initialization status. The following statuses are available:
+			 *
+			 *	* **unloaded**: the initial state; editable's instance has been created but
+			 *	is not fully loaded (in particular has no data),
+			 *	* **ready**: editable is fully initialized; `ready` status is set after
+			 *	first {@link CKEDITOR.editor#method-setData} has been called.
+			 *	* **detached**: the editable has been detached.
+			 *
+			 * @since 4.3.3
+			 * @readonly
+			 * @property {String}
+			 */
+			this.status = 'unloaded';
+
+			/**
 			 * Indicate whether the editable element has gained focus.
 			 *
 			 * @property {Boolean} hasFocus
@@ -97,9 +112,52 @@
 			},
 
 			/**
-			 * Registers an event listener that needs to be removed on detaching.
+			 * Registers an event listener that needs to be removed when detaching this editable.
+			 * This means that it will be automatically removed when {@link #detach} is executed,
+			 * for example on {@link CKEDITOR.editor#setMode changing editor mode} or destroying editor.
 			 *
-			 * @see CKEDITOR.event#on
+			 * Except for `obj` all other arguments have the same meaning as in {@link CKEDITOR.event#on}.
+			 *
+			 * This method is strongly related to the {@link CKEDITOR.editor#contentDom} and
+			 * {@link CKEDITOR.editor#contentDomUnload} events, because they are fired
+			 * when an editable is being attached and detached. Therefore, this method is usually used
+			 * in the following way:
+			 *
+			 *		editor.on( 'contentDom', function() {
+			 *			var editable = editor.editable();
+			 *			editable.attachListener( editable, 'mousedown', function() {
+			 *				// ...
+			 *			} );
+			 *		} );
+			 *
+			 * This code will attach the `mousedown` listener every time a new editable is attached
+			 * to the editor, which in classic (`iframe`-based) editor happens every time the
+			 * data or the mode is set. This listener will also be removed when that editable is detached.
+			 *
+			 * It is also possible to attach a listener to another object (e.g. to a document).
+			 *
+			 *		editor.on( 'contentDom', function() {
+			 *			editor.editable().attachListener( editor.document, 'mousedown', function() {
+			 *				// ...
+			 *			} );
+			 *		} );
+			 *
+			 * @param {CKEDITOR.event} obj The element/object to which the listener will be attached. Every object
+			 * which inherits from {@link CKEDITOR.event} may be used including {@link CKEDITOR.dom.element},
+			 * {@link CKEDITOR.dom.document}, and {@link CKEDITOR.editable}.
+			 * @param {String} eventName The name of the event that will be listened to.
+			 * @param {Function} listenerFunction The function listening to the
+			 * event. A single {@link CKEDITOR.eventInfo} object instance
+			 * containing all the event data is passed to this function.
+			 * @param {Object} [scopeObj] The object used to scope the listener
+			 * call (the `this` object). If omitted, the current object is used.
+			 * @param {Object} [listenerData] Data to be sent as the
+			 * {@link CKEDITOR.eventInfo#listenerData} when calling the listener.
+			 * @param {Number} [priority=10] The listener priority. Lower priority
+			 * listeners are called first. Listeners with the same priority
+			 * value are called in the registration order.
+			 * @returns {Object} An object containing the `removeListener`
+			 * function that can be used to remove the listener at any time.
 			 */
 			attachListener: function( obj, event, fn, scope, listenerData, priority ) {
 				!this._.listeners && ( this._.listeners = [] );
@@ -355,6 +413,11 @@
 					data = this.editor.dataProcessor.toHtml( data );
 
 				this.setHtml( data );
+
+				// Editable is ready after first setData.
+				if ( this.status == 'unloaded' )
+					this.status = 'ready';
+
 				this.editor.fire( 'dataReady' );
 			},
 
@@ -385,6 +448,8 @@
 			detach: function() {
 				// Cleanup the element.
 				this.removeClass( 'cke_editable' );
+
+				this.status = 'detached';
 
 				// Save the editor reference which will be lost after
 				// calling detach from super class.
@@ -1048,8 +1113,8 @@
 
 			var editable = editor.editable();
 
-			// Setup proper ARIA roles and properties for inline editable, framed
-			// editable is instead handled by plugin.
+			// Setup proper ARIA roles and properties for inline editable, classic
+			// (iframe-based) editable is instead handled by plugin.
 			if ( editable && editable.isInline() ) {
 
 				var ariaLabel = editor.title;